refactor(pipeline/api-direct-call): extract EnvelopeTypes from IApiDirectCallConfig
Phase 8 commit 4/8 of ApiDirectCallConfig split.

- Carve EnvelopeTypes (IFingerprintConfig, IJwtClaimsConfig, AuthScheme,
  IProbeConfig, IPreStepHook) from
  src/Scrapers/Pipeline/Mediator/ApiDirectCall/IApiDirectCallConfig.ts.
- New file:
  src/Scrapers/Pipeline/Mediator/ApiDirectCall/ConfigContracts/EnvelopeTypes.ts
  (56 raw LoC). Imports JsonValueTemplate from ./TemplateTypes.js,
  ICryptoFieldConfig from ./SignerTypes.js, plus WKQueryOperation /
  WKUrlGroup from ../../../Registry/WK — bucket-3 per the locked
  TemplateTypes/SignerTypes -> EnvelopeTypes chain.
- IApiDirectCallConfig.ts imports the five envelope types and continues
  to re-export them via its existing `export type {}` block so all 53
  callers compile unchanged. WKQueryOperation import dropped from the
  god-file (now owned by EnvelopeTypes); WKUrlGroup retained until
  Commit 5 carves IStepConfig out into FlowTypes.

Verified:
- npm run type-check green
- npx tsup green (lib/index.cjs byte-identical to main)
- npx jest --testPathPatterns=ApiDirectCall green (223/223)
- npx eslint src/Scrapers/Pipeline/Mediator/ApiDirectCall green
- New sub-module 56 LoC (well under 150 cap)

Master plan: pipeline-decoupling-master-2026-05-28 / phase-8
